01.init/requirements.txt
pillow==8.3.1
chainer==7.8.0
tensorflow==2.5.0
opencv-python==4.5.4.60